Optimize docker-compose setup
