--- docker-compose/config/talent/config.json
@@ -1,18 +1,0 @@
-{
-    "loglevel": "INFO",
-    "protocolGateway": {
-        "adapters": [
-            {
-                "platform": true,
-                "module": {
-                    "name": "./util/mqttClient",
-                    "class": "MqttProtocolAdapter"
-                },
-                "config": {
-                    "topicNamespace": "iotea/",
-                    "brokerUrl": "mqtt://mosquitto:1883"
-                }
-            }
-        ]
-    }
-}
