--- docker-compose/docker-compose.edge-test.yml
@@ -1,100 +1,0 @@
-version: '3.2'
-services:
-  mosquitto:
-    build:
-      context: ../iot-event-analytics/docker/mosquitto
-      dockerfile: Dockerfile.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    ports:
-      - '1883:1883'
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/mosquitto"
-        target: "/mosquitto/config"
-  iotea:
-    build:
-      context: ../iot-event-analytics
-      dockerfile: docker/platform/Dockerfile.slim.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/iotea"
-        target: "/app/docker/platform/config"
-    depends_on:
-      - mosquitto
-  kuksa.val:
-    image: ${KUKSA_VAL_IMG}
-    environment:
-      - KUKSAVAL_OPTARGS=--insecure
-    ports:
-      - '8090:8090'
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/kuksa.val"
-        target: "/config"
-  kuksa.val2iotea:
-    build:
-      context: ../iot-event-analytics
-      dockerfile: docker/kuksa.val2iotea/Dockerfile.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    depends_on:
-      - kuksa.val
-      - iotea
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/kuksa.val2iotea"
-        target: "/app/docker/kuksa.val2iotea/config"
-    depends_on:
-      - mosquitto
-      - kuksa.val
-  hal-interface-adapter:
-    build:
-      context: ../
-      dockerfile: src/edge.hal-interface-adapter/Dockerfile.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    depends_on:
-      - kuksa.val
-      - iotea
-      - hal-interface
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/hal-interface-adapter/config"
-        target: "/app/config"
-  hal-interface:
-    build:
-      context: ../
-      dockerfile: src/edge.hal-interface/Dockerfile.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    depends_on:
-      - mosquitto
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/hal-interface"
-        target: "/app/config"
-  talent:
-    build:
-      context: ../
-      dockerfile: src/edge.talent/Dockerfile.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    environment:
-      - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-      - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/talent"
-        target: "/app/config"
-    depends_on:
-      - iotea
-      - hal-interface-adapter
