+++ docker-compose/docker-compose.stack.yml
@@ -12,6 +12,8 @@ services:
         - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
     ports:
       - '1883:1883'
+    networks:
+      - edge-platform-network
     volumes:
       - type: bind
         source: "${CONFIG_DIR}/mosquitto"
@@ -26,42 +28,12 @@ services:
       args:
         - HTTP_PROXY=${DOCKER_HTTP_PROXY}
         - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
+    networks:
+      - edge-platform-network
     volumes:
       - type: bind
         source: "${CONFIG_DIR}/iotea"
         target: "/app/docker/platform/config"
-    depends_on:
-      - mosquitto
-  kuksa.val:
-    image: ${KUKSA_VAL_IMG}
-    environment:
-      - KUKSAVAL_OPTARGS=--insecure
-    ports:
-      - '8090:8090'
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/kuksa.val"
-        target: "/config"
-  kuksa.val2iotea:
-    platform: linux/${ARCH-amd64}
-    build:
-      context: ../iot-event-analytics
-      labels:
-        arch: ${ARCH-amd64}
-      dockerfile: ../iot-event-analytics/docker/kuksa.val2iotea/Dockerfile.${ARCH-amd64}
-      args:
-        - HTTP_PROXY=${DOCKER_HTTP_PROXY}
-        - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    depends_on:
-      - kuksa.val
-      - iotea
-    volumes:
-      - type: bind
-        source: "${CONFIG_DIR}/kuksa.val2iotea"
-        target: "/app/docker/kuksa.val2iotea/config"
-    depends_on:
-      - iotea
-      - kuksa.val
   hal-interface-adapter:
     platform: linux/${ARCH-amd64}
     build:
@@ -72,8 +44,8 @@ services:
       args:
         - HTTP_PROXY=${DOCKER_HTTP_PROXY}
         - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    depends_on:
-      - iotea
+    networks:
+      - edge-platform-network
     volumes:
       - type: bind
         source: "${CONFIG_DIR}/hal-interface-adapter/config-no-kuksa"
@@ -88,10 +60,13 @@ services:
       args:
         - HTTP_PROXY=${DOCKER_HTTP_PROXY}
         - HTTPS_PROXY=${DOCKER_HTTPS_PROXY}
-    depends_on:
-      - mosquitto
-      - hal-interface-adapter
+    networks:
+      - edge-platform-network
     volumes:
       - type: bind
         source: "${CONFIG_DIR}/hal-interface"
         target: "/app/config"
+networks:
+  edge-platform-network:
+    name: edge-platform-network
+    driver: bridge
