+++ docker-compose/run.properties
@@ -1,6 +1,5 @@
-USE_KUKSA_VAL=1
-USE_TALENT=1
+WITH_KUKSA_VAL=1
+WITH_TALENT=1
 DOCKER_IMAGE_BUILD=1
-DOCKER_IMAGE_EXPORT=1
+DOCKER_IMAGE_EXPORT=0
 DOCKER_CONTAINER_START=1
-IOTEA_VERSION=0.4.1
