+++ src/edge.hal-interface-adapter/.gitignore
@@ -1,1 +1,2 @@
 src/config
+*.tgz
