+++ src/edge.talent/.gitignore
@@ -1,1 +1,1 @@
-src/config
+*.tgz
