+++ src/edge.talent/src/index.js
@@ -8,51 +8,49 @@
  * SPDX-License-Identifier: MPL-2.0
  ****************************************************************************/
 
+const iotea = require('boschio.iotea');
 
-  const iotea = require('boschio.iotea');
+const Logger = iotea.util.Logger;
+process.env.LOG_LEVEL = Logger.ENV_LOG_LEVEL.INFO;
 
-  const Logger = iotea.util.Logger;
-  process.env.LOG_LEVEL = Logger.ENV_LOG_LEVEL.INFO;
+const {
+    MqttProtocolAdapter
+} = iotea.util;
 
-  const {
-      MqttProtocolAdapter
-  } = iotea.util;
+const {
+    Talent,
+    OpConstraint,
+    Rule,
+    AndRules,
+    ProtocolGateway,
+    TalentInput
+} = iotea;
 
-  const {
-      Talent,
-      OpConstraint,
-      Rule,
-      AndRules,
-      ProtocolGateway,
-      TalentInput
-  } = iotea;
+const {
+    VALUE_TYPE_RAW
+} = iotea.constants;
 
-  const {
-      VALUE_TYPE_RAW
-  } = iotea.constants;
+class SpeedTalent extends Talent {
+    constructor(protocolGatewayConfig) {
+        super('speed-talent', protocolGatewayConfig);
+        this.logger.info(`Created with config: ${JSON.stringify(protocolGatewayConfig)}`);
+    }
 
-  class SpeedTalent extends Talent {
+    getRules() {
+        return new AndRules([
+            new Rule(new OpConstraint('Speed', OpConstraint.OPS.ISSET, null, 'Vehicle', VALUE_TYPE_RAW))
+        ]);
+    }
 
-      constructor(protocolGatewayConfig) {
-          super('speed-talent', protocolGatewayConfig);
-          this.logger.info(`Created with config: ${JSON.stringify(protocolGatewayConfig)}`);
-      }
+    async onEvent(ev, evtctx) {
+        this.logger.info(`${ev.feature} : ${TalentInput.getRawValue(ev)}`);
+    }
+}
 
-      getRules() {
-          return new AndRules([
-              new Rule(new OpConstraint('Speed', OpConstraint.OPS.ISSET, null, 'Vehicle', VALUE_TYPE_RAW))
-          ]);
-      }
+// Update mqttAdapterConfig.config.brokerUrl, if you specified a different one in your configuration !!!
+const mqttAdapterConfig = MqttProtocolAdapter.createDefaultConfiguration();
+mqttAdapterConfig.config.brokerUrl = 'mqtt://mosquitto:1883'
 
-      async onEvent(ev, evtctx) {
-          this.logger.info(`${ev.feature} : ${TalentInput.getRawValue(ev)}`);
-      }
-  }
-
-  // Update mqttAdapterConfig.config.brokerUrl, if you specified a different one in your configuration !!!
-  const mqttAdapterConfig = MqttProtocolAdapter.createDefaultConfiguration();
-  mqttAdapterConfig.config.brokerUrl = 'mqtt://mosquitto:1883'
-
-  new SpeedTalent(
-      ProtocolGateway.createDefaultConfiguration([ mqttAdapterConfig ])
-  ).start();
+new SpeedTalent(
+    ProtocolGateway.createDefaultConfiguration([mqttAdapterConfig])
+).start();
